SR/SRW: fix compilation errors and warnings

+++ SR/SR_full_dos.c
@@ -1,6 +1,6 @@
 /**
  *
- *  Copyright (C) 2016 Roman Pauer
+ *  Copyright (C) 2016-2019 Roman Pauer
  *
  *  Permission is hereby granted, free of charge, to any person obtaining a copy of
  *  this software and associated documentation files (the "Software"), to deal in
@@ -136,18 +136,18 @@ int SR_disassemble_convert_fixup(const char *ostr, char *dst, fixup_data *fixup,
     }
     else if (fixup->type == FT_16BITOFS)
     {
-        sprintf(cbuf, "0x%x", (section[fixup->tsec].start + fixup->tofs) & 0xffff);
+        sprintf(cbuf, "0x%x", (unsigned int)((section[fixup->tsec].start + fixup->tofs) & 0xffff));
     }
     else
     {
-        sprintf(cbuf, "0x%x", section[fixup->tsec].start + fixup->tofs);
+        sprintf(cbuf, "0x%x", (unsigned int)(section[fixup->tsec].start + fixup->tofs));
     }
 
     // find substring (must be exactly one occurence)
     str1 = strstr(ostr, cbuf);
     if (str1 == NULL)
     {
-        fprintf(stderr, "Error: error converting fixup: %i - %s - %s (%i: 0x%x)\n", 1, ostr, cbuf, Entry, offset);
+        fprintf(stderr, "Error: error converting fixup: %i - %s - %s (%i: 0x%x)\n", 1, ostr, cbuf, Entry, (unsigned int)offset);
 
         return 1;
     }
@@ -198,7 +198,7 @@ int SR_disassemble_convert_fixup(const char *ostr, char *dst, fixup_data *fixup,
                 }
             }
         }
-        fprintf(stderr, "Two string occurences in fixup: 0x%x, %i - 0x%x\n%s\n%s\n", fixup->sofs, fixup->tsec, fixup->tofs, cbuf, ostr);
+        fprintf(stderr, "Two string occurences in fixup: 0x%x, %i - 0x%x\n%s\n%s\n", (unsigned int)fixup->sofs, (unsigned int)fixup->tsec, (unsigned int)fixup->tofs, cbuf, ostr);
         return 2;
     }
 
@@ -266,7 +266,7 @@ int SR_disassemble_convert_fixup(const char *ostr, char *dst, fixup_data *fixup,
     }
     else
     {
-        sprintf(&(dst[length1]), "(%s + (%i))", cbuf, fixup->tofs - (int_fast32_t)ofs);
+        sprintf(&(dst[length1]), "(%s + (%i))", cbuf, (int)(fixup->tofs - (int_fast32_t)ofs));
     }
 
     // copy second part of the string
@@ -311,7 +311,7 @@ int SR_disassemble_offset_dos(unsigned int Entry, uint_fast32_t offset)
 
         if (output == NULL)
         {
-            fprintf(stderr, "Error: output not found - %i - 0x%x\n", Entry, offset);
+            fprintf(stderr, "Error: output not found - %i - 0x%x\n", Entry, (unsigned int)offset);
 
             return 1;
         }
@@ -349,7 +349,7 @@ int SR_disassemble_offset_dos(unsigned int Entry, uint_fast32_t offset)
 
                 if (fixup->sofs > offset + decoded_length - fixuplen)
                 {
-                    fprintf(stderr, "Error: decoding fixup mismatch - %i - %i\n", Entry, offset);
+                    fprintf(stderr, "Error: decoding fixup mismatch - %i - %i\n", Entry, (unsigned int)offset);
 
                     return 2;
                 }
@@ -367,7 +367,7 @@ int SR_disassemble_offset_dos(unsigned int Entry, uint_fast32_t offset)
 
                             if (fixup2->sofs > offset + decoded_length - fixuplen2)
                             {
-                                fprintf(stderr, "Error: decoding fixup mismatch - %i - %i\n", Entry, offset);
+                                fprintf(stderr, "Error: decoding fixup mismatch - %i - %i\n", Entry, (unsigned int)offset);
 
                                 return 2;
                             }
@@ -646,7 +646,7 @@ int SR_disassemble_offset_dos(unsigned int Entry, uint_fast32_t offset)
                 break;
 
             case UD_Iinvalid:
-                printf("decoded length: %i (%i - 0x%x)\n", decoded_length, Entry, offset);
+                printf("decoded length: %i (%i - 0x%x)\n", decoded_length, Entry, (unsigned int)offset);
                 return 3;
                 break;
 
@@ -695,7 +695,7 @@ int SR_disassemble_offset_dos(unsigned int Entry, uint_fast32_t offset)
                         }
                         else
                         {
-                            printf("fixup: %i - 0x%x - 0x%x\n", fixup->tsec, fixup->tofs, section[fixup->tsec].start + fixup->tofs);
+                            printf("fixup: %i - 0x%x - 0x%x\n", (unsigned int)fixup->tsec, (unsigned int)fixup->tofs, (unsigned int)(section[fixup->tsec].start + fixup->tofs));
                             printf("op type: %i - %i - %i\n", ud_obj.operand[0].type, ud_obj.operand[1].type, ud_obj.operand[2].type);
                             printf("op size: %i - %i - %i\n", ud_obj.operand[0].size, ud_obj.operand[1].size, ud_obj.operand[2].size);
                             return -51;
